Put OpenRA.Test in front of the analyzer gate, and pin the vulnerable package out of the graph
Two build-system items in the same subsystem. Both were briefed from second-hand findings; both
premises were partly wrong, and the corrections changed what the right fix is.

`make check` on Linux/macOS compiles exactly two projects under -warnaserror: its Debug phase runs
`dotnet build` in the repo root, and that resolves WW3MOD.sln, whose whole project list is
OpenRA.Game and OpenRA.Mods.Common. The `check: engine` prerequisite builds the engine in Release,
where Directory.Build.props strips every analyzer. So adding the missing `.Build.0` to
engine/OpenRA.sln - the briefed fix - would NOT have gated OpenRA.Test here; it would only have
added it to the Release engine build, biting on Windows alone (make.ps1 routes through the engine's
own check target, which does build the engine solution under -warnaserror). That file is also
untouched since the 2023 engine import, i.e. upstream's choice, and under RUNTIME=mono the engine
builds as netstandard2.1, which Microsoft.NET.Test.Sdk does not target - a solution edit would hand
the mono lane a build it cannot do. Naming the .csproj in the check target's net6 branch is what
actually gates it, on both the Makefile and make.ps1 paths.

That exposed 47 unique violations across 8 rules, all in WW3MOD-authored fixtures. 32 are fixed
here, including two that were defects rather than style: a <paramref> on a type doc naming a
parameter that does not exist, and a foreach over Regex matches taking the non-generic IEnumerable
path. The remaining 15 are RCS1226 on the long multi-paragraph <summary> blocks that record which
regression each test guards; those are turned off by a scoped .editorconfig in the test project,
because the rule governs generated API documentation and OpenRA.Test publishes none. Every other
rule stays at full engine strength.

The vulnerable package is not NUnit's. NuGet.CommandLine 4.4.1 (GHSA-3885-8gqc-3wpf) is pulled by
rix0rrr.BeaconLib 1.0.2, referenced from OpenRA.Mods.Common - so it sits in the engine graph, and
updating NUnit.Console would have done nothing. BeaconLib 1.0.2 is its final release, so there is no
update path; a direct reference at a patched version with ExcludeAssets=All is what keeps the
vulnerable one out, and it costs nothing because the package contributes no compile, runtime or
build assets at all.

The audit is a target-framework effect, not an SDK effect. Measured on SDK 10.0.400 with a bare
restore: net6.0 and net8.0 do not raise NU1902, net10.0 does, because NuGetAuditMode only defaults
to `all` at net10.0. The global.json pin at 6.0.428 is therefore not what hides this - it surfaces
on the TFM bump commit itself, which is why clearing it first matters. The NU1510 "redundant"
System.* references are deliberately left alone: they are in-box at net6.0 but not at
netstandard2.1, so deleting them would break the mono lane.

RED controls for both zeros: the same sabotage that makes the new gate exit 2 with `error SA1210`
leaves the pre-change gate at exit 0, and reverting the scoped suppression brings back exactly the
30 raw RCS1226 lines and nothing else.

Verified on main @ 08b255f7: make check exits 0, make all clean, NUnit 1577/1577 (the 1569 in the
brief is a stale figure - this diff changes no test attributes).

+++ engine/OpenRA.Test/OpenRA.Mods.Common/SupplyProviderMathTest.cs
@@ -180,10 +180,10 @@ namespace OpenRA.Test
 		{
 			var totalSupply = 500;
 
-			Assert.That((float)500 / totalSupply, Is.EqualTo(1.0f));
-			Assert.That((float)250 / totalSupply, Is.EqualTo(0.5f));
-			Assert.That((float)0 / totalSupply, Is.EqualTo(0.0f));
-			Assert.That((float)100 / totalSupply, Is.EqualTo(0.2f));
+			Assert.That(500f / totalSupply, Is.EqualTo(1.0f));
+			Assert.That(250f / totalSupply, Is.EqualTo(0.5f));
+			Assert.That(0f / totalSupply, Is.EqualTo(0.0f));
+			Assert.That(100f / totalSupply, Is.EqualTo(0.2f));
 		}
 	}
 }
